line3(wip) and new function processor
cattyn function has little issues(the function has magenta color, look at the right part)

+++ src/main/java/the/kis/devs/g2i/FunctionHelper.java
@@ -6,8 +6,8 @@ package the.kis.devs.g2i;
  */
 public class FunctionHelper {
     public static FunctionArguments getLinearArguments(int x1, int y1, int x2, int y2) {
-        int a = (y2 - y1) / (x2 - x1);
-        int b = y1 - a * x1;
+        float a = (float) (y2 - y1) / (float) (x2 - x1);
+        float b = y1 - a *  x1;
 
         return new FunctionArguments(a, b);
     }
